Fix bug in MySQL connection pooling
If someone wants a connection and the pool of existing connections has
more than a single item, all items but the first will be lost.  This
eventually leads to a hang where the pool of connections is empty, but
the caching logic thinks the pool is full and won't create a new
connection.